refactor(cli): flatten github command module
Move the GitHub command slice to the neutral command root and update its internal export without changing public CLI or API behavior.
