+++ turbo/apps/cli/src/commands/github/upload-file.ts
@@ -4,8 +4,8 @@ import { Command } from "commander";
 import {
   completeGithubFileUpload,
   initGithubFileUpload,
-} from "../../../lib/api/domains/integrations-github";
-import { withErrorHandler } from "../../../lib/command/with-error-handler";
+} from "../../lib/api/domains/integrations-github";
+import { withErrorHandler } from "../../lib/command/with-error-handler";
 
 const MIME_BY_EXTENSION: Record<string, string> = {
   ".png": "image/png",
